backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
asyncpg>=0.28.0
psycopg2-binary>=2.9.7
pydantic>=2.3.0
python-multipart>=0.0.6
starlette>=0.27.0
websockets>=11.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.12.0
python-dotenv>=1.0.0